[Refactor] Move the Relay parts of an infinite scroll view into it's own file
