+++ src/lib/components/artist/artworks/index.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import { StyleSheet, View, ViewProperties } from "react-native"
 import * as Relay from "react-relay"
 
-import ArtworksGrid from "../../artwork_grids/infinite_scroll_grid"
+import ArtworksGrid from "../../artwork_grids/relay_connections/artist_artworks_grid"
 import Separator from "../../separator"
 import SerifText from "../../text/serif"
